Set scrum board column widths according to issue percentages
Allowing the columns to have widths set as percentages of the issues
show in the column from the total issues gives better results in terms
of layout.

There is still room for improvement, as the percentages may not be
multiples of the size of the scrum cards, but I don't see how this can
be done without using Javascript to detect the client's viewport
dimensions.

Tested on FireFox 6, Opera 11, Chromium 15 and IE 9 ( including
IE 8 mode ).

Fixes #222: Scrum board columns do not always fill all available space